+++ Held-Karp-algorithm/Held-Karp-algorithm/TSP.hpp
@@ -8,7 +8,6 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 #include <atomic>
 #include <chrono>
-#include <cstdint>
 #include <iostream>
 #include <sstream>
 #include <string>
@@ -24,21 +23,21 @@ namespace TSP
 	{
 	protected:
 		// precalculated
-		uint_least32_t POWER2[31] = { 1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024, 2048, 4096, 8192, 16384, 32768, 65536, 131072, 262144, 524288, 1048576, 2097152, 4194304, 8388608, 16777216, 33554432, 67108864, 134217728, 268435456, 536870912, 1073741824 };
+		unsigned int POWER2[31] = { 1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024, 2048, 4096, 8192, 16384, 32768, 65536, 131072, 262144, 524288, 1048576, 2097152, 4194304, 8388608, 16777216, 33554432, 67108864, 134217728, 268435456, 536870912, 1073741824 };
 
-		const vector<vector<uint_least16_t>> distance;
-		const uint_least16_t numberOfNodes;
+		const vector<vector<float>> distance;
+		const unsigned short numberOfNodes;
 
 		time_point<system_clock> begin;
-		atomic<uint_least16_t> currentCardinality = 2;
+		atomic<unsigned short> currentCardinality = 2;
 		atomic<bool> writingBuffer = false;
 
 	protected:
-		uint_least32_t Powered2Code(const vector<uint_least16_t> &S);
+		unsigned int Powered2Code(const vector<unsigned short> &S);
 
-		uint_least32_t Powered2Code(const vector<uint_least16_t> &S, const uint_least16_t exclude);
+		unsigned int Powered2Code(const vector<unsigned short> &S, const unsigned short exclude);
 
-		uint_least32_t Powered2Code(const uint_least32_t code, const uint_least16_t exclude);
+		unsigned int Powered2Code(const unsigned int code, const unsigned short exclude);
 
 		void ETL();
 		void ETLw();
@@ -46,14 +45,14 @@ namespace TSP
 		template <class T>
 		static T generateRandomNumber(const T startRange, const T endRange, const T limit);
 
-		virtual void Solve(uint_least16_t &opt, string &path) = 0;
+		virtual void Solve(float &opt, string &path) = 0;
 
 	public:
-		TSP(const vector<vector<uint_least16_t>> &DistanceMatrix2D);
+		TSP(const vector<vector<float>> &DistanceMatrix2D);
 
 		void Run();
 
-		static vector<vector<uint_least16_t>> New_RND_Distances(const uint_least16_t Size_of_RandomDistanceCosts);
+		static vector<vector<float>> New_RND_Distances(const unsigned short Size_of_RandomDistanceCosts);
 
 	};
 }
